+++ libtreebank.py
@@ -14,14 +14,110 @@ except ImportError:
     from ginger import libginger
 
 
+def trees_from_conll(lines_lst: ty.Iterable[str]) -> ty.Iterable[str]:
+    '''Extract individual tree strings from the lines of a CoNLL-like file.'''
+    current = []  # type: ty.List[str]
+    for line in lines_lst:
+        # Skip comment lines
+        if line.startswith('#'):
+            next
+        elif line.isspace():
+            if current:
+                yield current
+                current = []
+            next
+        else:
+            current.append(line)
+    # Flush the buffer at the end of the file
+    if current:
+        yield current
+
+
 # Generic formats
-def conllx(tree_str: str) -> libginger.Tree:
+def conllu(treebank_lst: ty.Iterable[str]) -> ty.Iterable[libginger.Tree]:
+    '''Parse a CoNLL-U treebank file and return its trees.'''
+    trees = trees_from_conll(treebank_lst)
+    return (_conllu_tree(t) for t in trees)
+
+
+def _conllu_tree(tree_str: ty.Iterable[str]) -> libginger.Tree:
+    """Return a tree (that is a list of `Nodes`) from its
+       [CoNNL-U string representation](http://universaldependencies.org/format.html)."""
+    root = libginger.Node(identifier=0, form='ROOT')
+    res = [root]
+    # First get the self-contained values, deal with references later
+    # IMPLEMENTATION: This relies on references being initialisable
+    #                 with identifiers
+    for i, line in enumerate(l.strip() for l in tree_str):
+        # Skip comment lines
+        if line.startswith('#'):
+            next
+
+        try:
+            identifier, form, lemma, upostag, xpostag, feats, head, deprel, deps, misc = line.split('\t')
+        except ValueError:
+            # TODO: Issue a warning here
+            raise libginger.ParsingError('At line {i} : 10 columns expected, got a {n} ({line!r})'.format(i=i, n=len(line.split('\t')), line=line))
+        try:
+            identifier = int(identifier)
+        except ValueError:
+            # TODO: Issue a warning here
+            if re.match(r'\d+-\d+', identifier):  # Skip multiword tokens
+                next
+            if re.match(r'\d+.\d+', identifier):  # Skip empty nodes
+                next
+            raise libginger.ParsingError('At line {i} : the `id` field does not respect CoNLL-U specifications.'.format(i=i))
+
+        try:
+            feats = dict() if feats == '_' else dict(e.split('=') for e in feats.split('|'))
+        except ValueError:
+            # Be nice : if empty, it should be an underscore, but let's be nice with spaces and empty strings, too
+            if feats.isspace() or not feats:
+                # TODO: Issue a warning here
+                feats = dict()
+            else:
+                raise libginger.ParsingError('At line {i} : the `feats` field does not respect CoNLL-U specifications.'.format(i=i))
+
+        try:
+            head = int(head)
+        except ValueError:
+            raise libginger.ParsingError('At line {i} : the `head` field does not respect CoNLL-U specifications.'.format(i=i))
+
+        try:
+            deps = [] if deps == '_' else [e.split(':') for e in deps.split('|')]
+        except ValueError as e:
+            # Be nice : if empty, it should be an underscore, but let's be nice with spaces and empty strings, too
+            if deps.isspace() or not deps:
+                # TODO: Add a warning here
+                deps = []
+            else:
+                raise libginger.ParsingError('At line {i} : the `deps` field does not respect CoNLL-U specifications.'.format(i=i))
+
+        new_node = libginger.Node(identifier, form, lemma, upostag, xpostag, feats, head, deprel, deps, misc)
+        res.append(new_node)
+
+    # Now deal with references, which is easy, since the index of a node in
+    # `res` is exactly its identifier
+    for n in res[1:]:
+        n.head = res[n.head]
+        n.deps = [(res[head], dep) for head, dep in n.deps]
+
+    return libginger.Tree(res)
+
+
+def conllx(treebank_lst: ty.Iterable[str]) -> ty.Iterable[libginger.Tree]:
+    '''Parse a CoNLL-X treebank file and return its trees.'''
+    trees = trees_from_conll(treebank_lst)
+    return (_conllx_tree(t) for t in trees)
+
+
+def _conllx_tree(tree_lst: ty.Iterable[str]) -> libginger.Tree:
     """Create an Universal Dependencies tree from a CoNLL-X tree."""
     root = libginger.Node(identifier=0, form='ROOT')
     res = [root]
     conllx_to_conllu_identifers = {0: 0}
 
-    for i, line in enumerate(l.strip() for l in tree_str.splitlines()):
+    for i, line in enumerate(l.strip() for l in tree_lst):
         # Skip comment lines
         if line.startswith('#'):
             next
@@ -85,7 +181,13 @@ def conllx(tree_str: str) -> libginger.Tree:
     return libginger.Tree(res)
 
 
-def conll2009_gold(tree_str: str) -> libginger.Tree:
+def conll2009_gold(treebank_lst: ty.Iterable[str]) -> ty.Iterable[libginger.Tree]:
+    '''Parse a CoNLL-2009 gold treebank file and return its trees.'''
+    trees = trees_from_conll(treebank_lst)
+    return (_conll2009_gold_tree(t) for t in trees)
+
+
+def _conll2009_gold_tree(tree_lst: ty.Iterable[str]) -> libginger.Tree:
     """Create an Universal Dependencies tree from a CoNLL-2009 tree.
        This takes only the gold columns into account
 
@@ -94,7 +196,7 @@ def conll2009_gold(tree_str: str) -> libginger.Tree:
     res = [root]
     conllx_to_conllu_identifers = {0: 0}
 
-    for i, line in enumerate(l.strip() for l in tree_str.splitlines()):
+    for i, line in enumerate(l.strip() for l in tree_lst):
         # Skip comment lines
         if line.startswith('#'):
             next
@@ -185,7 +287,13 @@ def to_conll2009_gold(tree: libginger.Tree) -> str:
     return '\n'.join(_node_to_conll2009_gold(n) for n in tree.nodes[1:])
 
 
-def conll2009_sys(tree_str: str) -> libginger.Tree:
+def conll2009_sys(treebank_lst: ty.Iterable[str]) -> ty.Iterable[libginger.Tree]:
+    '''Parse a CoNLL-U treebank file and return its trees.'''
+    trees = trees_from_conll(treebank_lst)
+    return (_conll2009_sys_tree(t) for t in trees)
+
+
+def _conll2009_sys_tree(tree_lst: ty.Iterable[str]) -> libginger.Tree:
     """Create an Universal Dependencies tree from a CoNLL-2009 tree.
        This takes only the predicted columns into account
 
@@ -194,7 +302,7 @@ def conll2009_sys(tree_str: str) -> libginger.Tree:
     res = [root]
     conllx_to_conllu_identifers = {0: 0}
 
-    for i, line in enumerate(l.strip() for l in tree_str.splitlines()):
+    for i, line in enumerate(l.strip() for l in tree_lst):
         # Skip comment lines
         if line.startswith('#'):
             next
@@ -216,7 +324,7 @@ def conll2009_sys(tree_str: str) -> libginger.Tree:
             feat = dict() if pfeat == '_' else dict(e.split('=') for e in pfeat.split('|'))
         except ValueError:
             # Be nice : if empty, it should be an underscore, but let's be nice with spaces and empty strings, too
-            if feat.isspace() or not feat:
+            if pfeat.isspace() or not pfeat:
                 # TODO: Issue a warning here
                 feat = dict()
             else:
@@ -229,7 +337,7 @@ def conll2009_sys(tree_str: str) -> libginger.Tree:
         try:
             head = int(head)
         except ValueError:
-            if head == '_' and pdeprel == '_':
+            if head == '_' and deprel == '_':
                 head, deprel = None, None
             else:
                 raise libginger.ParsingError('At line {i} : the `head` field does not respect CoNLL-2009 specifications.'.format(i=i))
@@ -281,12 +389,7 @@ def _node_to_conll2009_sys(node: libginger.Node):
 
 def to_conll2009_sys(tree: libginger.Tree) -> str:
     '''Return a CoNLL-2009 representation of the tree.'''
-    return '\n'.join(_node_to_conll2009_gold(n) for n in tree.nodes[1:])
-
-
-def conllu(tree_str: str) -> libginger.Tree:
-    """Create an Universal Dependencies tree from a CoNLL-U tree."""
-    return libginger.Tree.from_conll(tree_str)
+    return '\n'.join(_node_to_conll2009_sys(n) for n in tree.nodes[1:])
 
 
 def to_conllu(tree: libginger.Tree) -> str:
@@ -295,13 +398,19 @@ def to_conllu(tree: libginger.Tree) -> str:
 
 
 # Parser-specific formats
-def talismane(tree_str: str) -> libginger.Tree:
+def talismane(treebank_lst: ty.Iterable[str]) -> ty.Iterable[libginger.Tree]:
+    '''Parse a CoNLL-U treebank file and return its trees.'''
+    trees = trees_from_conll(treebank_lst)
+    return (_talismane_tree(t) for t in trees)
+
+
+def _talismane_tree(tree_str: ty.Iterable[str]) -> libginger.Tree:
     """Create an Universal Dependencies tree from a Talismane tree.
 
        Talismane outputs are essentially CoNLL-X files, with incompatible
        stylistic idiosyncrasies."""
-    conllx_str = re.sub(r'\|\t', r'\t', tree_str)
-    return conllx(conllx_str)
+    conllx_str = (s.replace(r'\|\t', r'\t') for s in tree_str)
+    return _conllx_tree(conllx_str)
 
 
 def dict_to_conll_map(d: ty.Dict) -> str:
@@ -310,10 +419,10 @@ def dict_to_conll_map(d: ty.Dict) -> str:
 
 
 # Guessing tools
-def guess(filecontents: str) -> str:
+def guess(filelines: ty.Iterable[str]) -> str:
     """Guess the format of a file. Return the name of the format in a way
        as a key in `formats`."""
-    lines = filecontents.split('\n')
+    lines = iter(filelines)
     first_line_columns = next(l for l in lines if l).split('\t')
 
     if len(first_line_columns) == 10:  # 10 columns, assuming CoNLL-[XU] of some kind
@@ -323,7 +432,7 @@ def guess(filecontents: str) -> str:
                 return 'talismane'
             return 'conllx'
     elif len(first_line_columns) >= 14:  # 14 columns or more assume CoNLL-2009:
-        if any(l.split('\t')[2] != '_' for l in lines if l and not l.isspace()):
+        if any(l.split('\t')[3] != '_' for l in lines if l and not l.isspace()):
             return 'conll2009_sys'
         return 'conll2009_gold'